更新数据库文档：t_it_bill_detail商品数量改为decimal(19,8)

+++ doc/99 SQL/01CreateTables.sql
@@ -538,7 +538,7 @@ CREATE TABLE IF NOT EXISTS `t_it_bill_detail` (
   `id` varchar(255) NOT NULL,
   `date_created` datetime DEFAULT NULL,
   `goods_id` varchar(255) NOT NULL,
-  `goods_count` int(11) NOT NULL,
+  `goods_count` decimal(19,8) NOT NULL,
   `show_order` int(11) NOT NULL,
   `itbill_id` varchar(255) NOT NULL,
   `data_org` varchar(255) DEFAULT NULL,
